apps/api/requirements.txt
# Rabbit Hole API — v0 minimal surface
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
httpx>=0.26.0
Pillow>=10.0.0
pytesseract>=0.3.10
openai>=1.0.0
